refactored moved picture function
