+++ tutorials/tutorial_performancesettings.py
@@ -31,7 +31,7 @@ import shutil
 from sys import platform
 from pyscreeze import ImageNotFoundException
 from tutorials.utils import platform_keys, start, finish, create_tutorial_images
-from tutorials.pictures import picture
+from tutorials.utils.picture import picture
 
 
 def automate_performance():
