+++ tutorials/tutorial_satellitetrack.py
@@ -28,7 +28,7 @@ import os.path
 from sys import platform
 from pyscreeze import ImageNotFoundException
 from tutorials.utils import platform_keys, start, finish, create_tutorial_images
-from tutorials.pictures import picture
+from tutorials.utils.picture import picture
 
 
 def automate_rs():
