+++ tutorials/tutorial_views.py
@@ -28,7 +28,7 @@ import pyautogui as pag
 from sys import platform
 from pyscreeze import ImageNotFoundException
 from tutorials.utils import platform_keys, start, finish, create_tutorial_images, get_region
-from tutorials.pictures import picture
+from tutorials.utils.picture import picture
 
 
 # ToDo in sideview and topview waypoint movement needs adjustment
